+++ package.json
@@ -25,14 +25,10 @@
     "express": "~4.15.2",
     "express-session": "^1.15.3",
     "find-cache-dir": "^1.0.0",
-    "fullcalendar": "^3.4.0",
     "google-auth-library": "^0.10.0",
     "google-calendar": "^1.3.2",
     "googleapis": "^20.0.1",
-    "jquery": "^3.2.1",
-    "jsdom": "^11.1.0",
     "jstz": "^1.0.9",
-    "moment": "^2.18.1",
     "mongoose": "^4.10.8",
     "morgan": "~1.8.1",
     "passport": "^0.3.2",
@@ -43,9 +39,7 @@
     "react-modal": "^2.2.1",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
-    "serve-favicon": "~2.4.2",
-    "vex-dialog": "^1.0.7",
-    "vex-js": "^4.0.0"
+    "serve-favicon": "~2.4.2"
   },
   "devDependencies": {
     "babel-core": "^6.25.0",
@@ -54,11 +48,9 @@
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.4",
     "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.2",
     "html-webpack-plugin": "^2.29.0",
     "nodemon": "^1.11.0",
     "style-loader": "^0.18.2",
-    "url-loader": "^0.5.9",
     "webpack": "^2.2.0",
     "webpack-dev-server": "^2.5.0"
   }
