+++ src/client/app/App.jsx
@@ -7,7 +7,6 @@ import Home from './components/Home';
 import  NotFound from './components/NotFound';
 import Profile from './components/Profile';
 import Add from './components/Add';
-import Calendar from './components/Calendar';
 
 import '../index.css';
