+++ src/server/routes/index.js
@@ -11,7 +11,7 @@ const router = express.Router();
 
 // GET home page.
 router.get('/', (req, res, next) => {
-  res.sendFile(path.join(__dirname, '../../../public/index.html'));
+  res.sendFile(path.join(__dirname, '../../../public'));
 });
 
 // Consildated routes from individual files.
